設定画面のリファクタリング: 階層構造と自動同期
- メイン設定画面をシンプルなリスト表示に変更
- 秘密鍵管理、リレーサーバー管理、アプリ設定を別画面に分離
- GoRouterに新しいルート追加 (/settings/secret-key, /settings/relays, /settings/app)
- 手動同期ボタンを削除し、完全自動バックグラウンド同期に変更
- 非推奨のwithOpacityをwithValuesに修正
- OS設定画面風のUIデザインを採用
